Move tag selection into the ReportContext
