+++ src/Libraries/Microsoft.Extensions.AI.Evaluation.Reporting/TypeScript/components/TagsDisplay.tsx
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 import { makeStyles, tokens } from '@fluentui/react-components';
+import { useReportContext } from './ReportContext';
 
 const useStyles = makeStyles({
   tagsContainer: {
@@ -112,12 +113,11 @@ export function GlobalTagsDisplay({ globalTags }: { globalTags: TagInfo[] }) {
 
 export interface FilterableTagsDisplayProps {
   filterableTags: TagInfo[];
-  onTagClick: (tag: string) => void;
-  selectedTags: string[];
 }
 
-export function FilterableTagsDisplay({ filterableTags, onTagClick, selectedTags }: FilterableTagsDisplayProps) {
+export function FilterableTagsDisplay({ filterableTags }: FilterableTagsDisplayProps) {
   const classes = useStyles();
+  const {selectedTags, handleTagClick} = useReportContext();
 
   if (filterableTags.length === 0) {
     return null;
@@ -133,7 +133,7 @@ export function FilterableTagsDisplay({ filterableTags, onTagClick, selectedTags
             key={tag} 
             className={`${classes.tagBubble} ${isSelected(tag) ? 'selected' : ''}`}
             title={`${tag} (appears on ${count} results) - Click to filter by this tag`}
-            onClick={() => onTagClick(tag)}
+            onClick={() => handleTagClick(tag)}
           >
             {tag}
           </div>
